Phase 1a: extract pure transpile logic + browser Faust editor scaffolding
Splits the Faust→AS transpiler into a pure ESM core that runs in both
Node and the browser, and adds the in-app Faust editor UI shell.

* New: wasmaudioworklet/faust/transpile-core.js (~1400 lines)
  Pure reshape/codegen. No fs/path/child_process. Exports: toClassName,
  extractUIFromJSON, parseASSource, parseComputeBody, reshapeASLine,
  reshapeInstanceConstants, reshapeInstanceClear, reshapeSIGInit,
  transpileDsp, transpileMastering, generateNRPNSetParam,
  generateNRPNDefaults, assembleSingleFile, assembleBundle.
  process.exit(...) calls become throw new Error(...). Path-based
  identifiers (sourceFile, basename) come in as parameters.

* tools/faust2as/faust2asc.js: 1551 → 232 lines. Now just a Node CLI
  that reads .dsp files, mounts sibling .lib/.dsp into the libfaust
  virtual FS, runs faust, then delegates to transpile-core. Verified
  master_me regen is byte-identical and dx7 ASC bundle's
  auto-generated portion is byte-identical (only the documented
  manual drum-kit section differs).

* New: wasmaudioworklet/faust/browser-transpile.js
  Lazy-loads @psalomo/faustwasm from jsDelivr, mounts sibling files,
  detects voice vs stereo-mastering via getNumInputs/Outputs, calls
  the appropriate transpile-core entry. Returns assembled .ts string.

* wasmaudioworklet/app.html.js: new Faust editor toggle alongside
  song/synth/presets/shader plus the editor area (file dropdown,
  "New file" input, save status, CodeMirror mount point).

* wasmaudioworklet/wasmgit/{wasmgitclient,wasmgitworker}.js: new
  listfiles(prefix) API that walks the working tree (so just-staged
  files appear too) and returns paths matching an optional prefix.
  The Faust editor uses this to populate the .dsp dropdown.
